- renamed functions in scalar to remove the n_ prefix, renamed few functions in vec3/4/quat to avoid ambiguous calls - templated some of the basic functions in scalar (min,max,clamp) - streamsizes are not 64bit based - fixed a pile of unclean 64bit handling, added asserts when possible - made warnings a bit more aggressive, added permissive- to windows flags. almost possible to compile with highest warning level in windows

+++ tests/testjobs/jobstest.cc
@@ -37,7 +37,7 @@ JobsTest::Run()
     CreateJobPortInfo portInfo;
     portInfo.name = "TestJobPort";
     portInfo.numThreads = 2;
-    portInfo.priority = FLT_MAX;
+    portInfo.priority = UINT_MAX;
     JobPortId port = CreateJobPort(portInfo);
 
     // create a job sync
